functions/requirements.txt
firebase_functions~=0.1.0
google.cloud.firestore~=2.12.0
google-cloud-vision~=3.4.4
langchain~=0.0.300
openai~=0.28.0
python-dotenv~=1.0.0
google-cloud-aiplatform~=1.33.1
transformers~=4.33.3